% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{house_prices}
\alias{house_prices}
\title{A historical dataset on nominal annual house prices from \cite{Knoll2017}
and respective macroeconomic variables from \cite{Jorda2017}.}
\format{
A matrix with 984 rows and 18 columns. Each column corresponds to
either the ISO/name of the country and year or one of the measurements
described above recorded for that combination of country-year.
}
\source{
\url{https://www.openicpsr.org/openicpsr/project/113055/version/V1/view}
        \url{https://www.macrohistory.net/database/}
}
\usage{
data("house_prices")
}
\description{
The dataset consists of nominal house prices and other macroeconomic
variables for 8 advanced economies over the time period 1890-2012.
}
\details{
The data are distributed through openICPSR, a public access repository
supported by the Inter-university Consortium for
Political and Social Research (ICPSR), under an Other License.

The dataset is freely available and was downloaded on 13 January 2022 from
https://www.openicpsr.org/openicpsr/project/113055/version/V1/view.

The macroeconomic variables are obtained via
the Jordà-Schularick-Taylor Macrohistory Database (\cite{Jorda2017}).
The Jordà-Schularick-Taylor Macrohistory Database is freely available and
may be used, shared and adapted subject to Creative Commons
Attribution-NonCommercial-ShareAlike 4.0 International (CC BY-NC-SA 4.0).

The data were downloaded on 13 January 2022 from
https://www.macrohistory.net/database/.

The entries in the dataset present one of the following measurements:
consumer price index (\verb|cpi|), real GDP (\verb|rgdppc|),
population size (\verb|pop|), long-term interest rate (\verb|ltrate|),
nominal house prices (\verb|hpnom|), real house prices (\verb|hpreal|)
obtained through deflation the nominal house prices with the corresponding
consumer price index (CPI), inflation (\verb|infl|) which is measured as
change in CPI, logarithms of real house prices, population size and real GDP
(\verb|log_hp|, \verb|log_gdp| and \verb|log_pop|, respectively) and
the changes of the log house prices, log population size, log real GDP,
long-term interest rate and inflation (\verb|delta_log_hp|,
\verb|delta_log_gdp|, \verb|delta_log_pop|, \verb|delta_ltrate|,
and \verb|delta_infl|, respectively). All of the measurements are recorded
per country (columns \verb|iso| and \verb|country|) and per year
(column \verb|year|).
}
\keyword{datasets}
