% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{house_prices}
\alias{house_prices}
\title{House price indices and macro-financial indicators for advanced economies
per country.}
\format{
A matrix with 2214 rows and 18 variables:
\itemize{
  \item iso: ISO code of the country
  \item year: year of the observation
  \item country: name of the country
  \item cpi: consumer price index (CPI) from the Jordà-Schularick-Taylor Macrohistory Database
  \item rgdppc: real GDP from the Jordà-Schularick-Taylor Macrohistory Database
  \item pop: population size from the Jordà-Schularick-Taylor Macrohistory Database
  \item ltrate: long-term interest rate from the Jordà-Schularick-Taylor Macrohistory Database
  \item hpnom: nominal house prices
  \item hpreal: deflated nominal house prices with the CPI
  \item infl: inflation that is measured as change in CPI
  \item log_hp: logarithm of the real house prices
  \item log_gdp: logarithm of the real GDP
  \item log_pop: logarithm of the population size
  \item delta_log_hp: first difference of the logarithm of the real house prices
  \item delta_log_gdp: first difference of the logarithm of the real GDP
  \item delta_log_pop: first difference of the logarithm of the population size
  \item delta_ltrate: first difference of the long-term interest rate
  \item delta_infl: first difference of the inflation
 }
}
\source{
\url{https://www.openicpsr.org/openicpsr/project/113055/version/V1/view},
\url{https://www.macrohistory.net/database/}
}
\usage{
data("house_prices")
}
\description{
Data that combines residential house price indices  for 14 advanced economies
over the years 1870 to 2012 from Knoll et al. (2017) with several
macro-financial variables from the Jordà-Schularick-Taylor Macrohistory Database.
}
\details{
We .

We also include the logarithms of real house prices (log_hp), the real GDP (log_gdp)
and the population size (log_pop) to measure the growth rates of these variables.

Furthermore, we include the first difference of logarithm of real house
prices (delta_log_hp), the logarithm of the real GDP (delta_log_gdp),
the logarithm of the population size (delta_log_pop), long-term interest
rate (delta_ltrate) and inflation (delta_infl) in order to simplify
further calculations.
}
\keyword{datasets}
